# backend/src/main/resources/openapi.yaml
openapi: 3.0.0

info:
  title: Reach Backend API
  version: 0.0.1-SNAPSHOT
  description: |
    This is the API specification for the endpoints included in the Reach web app backend.

paths:
  /health:
    get:
      tags:
        - Health
      summary: Ping backend health
      description: Call this endpoint to check if the backend is up and running.
      operationId: checkHealth
      responses:
        '200':
          description: Backend is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloWorld'
        '401':
          description: Health error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloWorldError'
        '500':
          description: Internal server error
  /brand:
    post:
      tags:
        - Brand
      summary: Add a new Brand
      description: Add a new Brand to the database
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
      responses:
        '201':
          description: Brand successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
    put:
      tags:
        - Brand
      summary: Update an existing Brand
      description: Update an existing Brand in the database
      operationId: updateBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
      responses:
        '204':
          description: Brand successfully updated
        '404':
          description: Brand to be updated not found
  /brand/{id}:
    get:
      tags:
        - Brand
      summary: Get an existing Brand
      description: Get an existing Brand by its id
      operationId: getBrand
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the brand to retrieve
      responses:
        '200':
          description: Brand successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '404':
          description: Brand not found

  /creator:
    post:
      tags:
        - Creator
      summary: Create a new creator
      description: Creates a new creator with the given details.
      operationId: createCreator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Creator'
      responses:
        '201':
          description: Creator successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Creator'
        '400':
          description: Invalid input
    put:
      tags:
        - Creator
      summary: Update an existing creator
      operationId: updateCreator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Creator'
      responses:
        '204':
          description: Creator updated
        '404':
          description: Creator not found

  /creator/{id}:
    get:
      tags:
        - Creator
      summary: Get creator by ID
      operationId: getCreator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Creator found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Creator'
        '404':
          description: Creator not found

components:
  schemas:
    HelloWorld:
      type: object
      description: Default health response
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    HelloWorldError:
      type: object
      description: Health error
      properties:
        error:
          type: string
        randomNumber:
          type: number
    Brand:
      type: object
      description: Brand created on the frontend level containing all details
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: string
          format: uri
        revenuePerMonth:
          type: integer
        itemsSold:
          type: integer
        averageUnitPrice:
          type: number
          format: float
        niches:
          type: array
          items:
            $ref: '#/components/schemas/Niche'
        region:
          type: string
    Creator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        salesPerMonth:
          type: integer
        itemsSold:
          type: integer
        niches:
          type: array
          items:
            $ref: '#/components/schemas/Niche'
        region:
          type: string

    Niche:
      type: string
      enum:
        - BEAUTY_AND_PERSONAL_CARE
        - HEALTH
        - NUTRITION_AND_WELLNESS
        - SPORTS_AND_OUTDOOR
        - VITAMINS_MINERALS_AND_WELLNESS_SUPPLEMENTS
        - PHONES_AND_ELECTRONICS
        - FASHION_ACCESSORIES
        - FRAGRANCE
        - SKINCARE
      description: Enum representing the niche of the creator.